While working PR 403, I noticed these updates needed in personalization
- PersonalizationService was still caching the server base url. That was
previously moved to the DiscoveryService.
- Personalization still had the method to calculate the web socket url.
Also previously moved to DiscoveryService.
- There were several BehaviorSubjects in PersonalizationService used for
notifying observers of the URL changes.  I moved these to
DiscoveryService since DiscoveryService handles calculation of the URLs.
- So that the DiscoveryService could observe changes to the base server
url attributes and in turn inform its observers of URL changes, I added
a couple of BehaviorSubjects to the PersonalizationService.
- Updated all dependent unit tests.
- Unrelated update: add  "core-js": "^2.5.7" dependency to
openpos-client-personalize project so that it would build