Fix managment server connection and startup failed retry
- Found an issue at a retailer where the OpenPOS management server was
returning the web socket URL for a client to connect before the
websocket could actually be established on the server side. Presumably
due to the OpenPOS server 'status' web service being available before
the websocket infrastructure was initialized and ready for a client
connection.  The client already had some handling for this in the
subscribe-to-session-task.ts by allowing up to a default of 7500ms for
the websocket session to be established.  That wasn't enough time,
however in all cases.  To fix this, I added Configuration parameter in
the client Configuration to control how long we wait for the web socket
connection to be confirmed and I have set the value to 20000ms.
- A second issue was reported that in the case when the startup failed
(due to the connection not being able to be confirmed within the
allotted time), the 'Retry' button was resulting in showing the white
screen of death. This was due to us doing a location.reload() from the
startup failed component.  At that point, however, the location was the
new route (the startup failed route) and so reloading that rewritten url
would cause the white screen.  I made a change to save off the current
location.href before running any of the startup tasks so that the
initial location could be reloaded upon failure.  This ensures that any
URL parameters, etc. are also passed through on retry.