Work around bug with Jackson that allows errant convert of map to Form
- A retailer found that if a screen is sent from server to client that
contains a form, but the screen sends back an alternate action that
doesn't contain a form, the state that processes the action receives an
action whose getData() method returns an empty Form object.
- Upon debugging, it turns out that Jackson was converting the returned
LinkedHashMap of data on the action to an empty Form object even though
the attributes in the map were not assignment compatible with a form.
- I've added a Form.isAssignableFrom static method to check whether or
not a given object can be assigned to a Form object.  I also added unit
tests to check and guard behavior.